build: Create gtest for 'tv_device' sample program
To get this to run 'upnpapi.cpp' needs to be reworked to fit exactly to
the forked pupnp version 1.14.12.

+++ pupnp/upnp/src/threadutil/ThreadPool.cpp
@@ -405,8 +405,8 @@ static void SetSeed(void) {
 
     gettimeofday(&t, NULL);
 #if defined(__PTW32_DLLPORT)
-    srand((unsigned int)t.tv_usec +
-          (unsigned int)ithread_get_current_thread_id().p);
+    srand((unsigned long long)t.tv_usec +
+          (unsigned long long)ithread_get_current_thread_id().p);
 #elif defined(BSD) || defined(__APPLE__) || defined(__FreeBSD_kernel__)
     srand((unsigned long)t.tv_usec +
           (unsigned long)ithread_get_current_thread_id());
